(NOT WORKING) benchmark push
Isolated problem to somewhere between lines 55-93. The issue is that it's marking some words as not possible despite them being possible(i.e., if statements being true when they're not supposed to be). Left lots of commented out print statements that might help, but there's too much data to manually comb.